requirements.txt
selenium>=3.8.1
behave>=1.2
Appium-Python-Client>=0.26